JSON TranslateNumberToGerman
If askForNumber returns a number between 1 and 10, translateNumberToGerman returns the German translation of that number as a string.

If askForNumber does not return a number between 1 and 10 and instead throws an exception, translateNumberToGerman logs the error message to the console and restarts the whole process. The user should keep being prompted to input a number between 1 and 10 until she does so.