Twitter API Part2
Promisify getToken and getTweets process and use Promise.all to add more two news source. The requests for tweets from all three news sources was made simultaneously. The combined list of tweets that we send to the client were sorted in reverse chronological order. Appending to each tweet the name of the source in parentheses.